Moved job explorer specific components under the correct folder
